+++ app/src/main/java/app/revanced/integrations/shared/Utils.java
@@ -37,6 +37,9 @@ import java.util.concurrent.TimeUnit;
 import app.revanced.integrations.shared.settings.BooleanSetting;
 import kotlin.text.Regex;
 
+import app.revanced.integrations.twitter.settings.SettingsStatus;
+import app.revanced.integrations.twitter.patches.FeatureSwitchPatch;
+
 public class Utils {
 
     @SuppressLint("StaticFieldLeak")
@@ -46,6 +49,11 @@ public class Utils {
 
     private Utils() {
     } // utility class
+
+    public static void load() {
+        SettingsStatus.load();
+        FeatureSwitchPatch.load();
+    }
 
     public static String getVersionName() {
         if (versionName != null) return versionName;
@@ -172,6 +180,10 @@ public class Utils {
      */
     public static int getResourceIdentifier(@NonNull String resourceIdentifierName, @NonNull String type) {
         return getResourceIdentifier(getContext(), resourceIdentifierName, type);
+    }
+
+    public static String getResourceString(@NonNull String resourceIdentifierName) throws Resources.NotFoundException {
+        return getContext().getResources().getString(getResourceIdentifier(resourceIdentifierName, "string"));
     }
 
     public static int getResourceInteger(@NonNull String resourceIdentifierName) throws Resources.NotFoundException {