+++ app/src/main/java/app/revanced/integrations/twitter/settings/SettingsActivity.java
@@ -10,26 +10,43 @@ import androidx.appcompat.widget.Toolbar;
 import app.revanced.integrations.shared.Utils;
 import app.revanced.integrations.shared.settings.BooleanSetting;
 import app.revanced.integrations.shared.settings.StringSetting;
+import app.revanced.integrations.twitter.settings.featureflags.FeatureFlagsFragment;
 import com.twitter.ui.widget.LegacyTwitterPreferenceCategory;
-
 @SuppressWarnings("deprecation")
 public class SettingsActivity extends Activity {
+    public static Toolbar toolbar;
+
     @SuppressLint("ResourceType")
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
         setContentView(Utils.getResourceIdentifier("preference_fragment_activity", "layout"));
-        Toolbar toolbar = findViewById(Utils.getResourceIdentifier("toolbar", "id"));
+        toolbar = findViewById(Utils.getResourceIdentifier("toolbar", "id"));
         toolbar.setNavigationIcon(Utils.getResourceIdentifier("ic_vector_arrow_left", "drawable"));
-        toolbar.setTitle("Mod Settings");
+        toolbar.setTitle(Utils.getResourceString("piko_title_settings"));
         toolbar.setNavigationOnClickListener(view -> onBackPressed());
 
-        getFragmentManager().beginTransaction().add(Utils.getResourceIdentifier("fragment_container", "id"), new Screen()).commit();
+        getFragmentManager().beginTransaction().replace(Utils.getResourceIdentifier("fragment_container", "id"), new Screen()).commit();
     }
 
-    public static class Screen extends PreferenceFragment implements Preference.OnPreferenceClickListener{
+    @Override
+    public void onBackPressed() {
+        if (getFragmentManager().getBackStackEntryCount()>0) {
+            getFragmentManager().popBackStack();
+        } else {
+            super.onBackPressed();
+        }
+    }
+
+    public static class Screen extends PreferenceFragment implements Preference.OnPreferenceClickListener {
         private Context context;
+
+        @Override
+        public void onResume() {
+            super.onResume();
+            SettingsActivity.toolbar.setTitle(Utils.getResourceString("piko_title_settings"));
+        }
 
         @Override
         public void onCreate(@Nullable Bundle savedInstanceState) {
@@ -39,12 +56,12 @@ public class SettingsActivity extends Activity {
             PreferenceScreen screen = getPreferenceManager().createPreferenceScreen(context);
 
             if (SettingsStatus.enablePremiumSection()) {
-                LegacyTwitterPreferenceCategory premiumPrefs = preferenceCategory("Premium", screen);
+                LegacyTwitterPreferenceCategory premiumPrefs = preferenceCategory(Utils.getResourceString("piko_title_premium"), screen);
                 if (SettingsStatus.enableReaderMode) {
                     premiumPrefs.addPreference(
                             switchPreference(
-                                    "Enable reader mode",
-                                    "Enables \"reader mode\" on long threads",
+                                    Utils.getResourceString("piko_pref_reader_mode"),
+                                    Utils.getResourceString("piko_pref_reader_mode_desc"),
                                     Settings.PREMIUM_READER_MODE
                             )
                     );
@@ -52,15 +69,15 @@ public class SettingsActivity extends Activity {
                 if (SettingsStatus.enableUndoPosts) {
                     premiumPrefs.addPreference(
                             switchPreference(
-                                    "Enable undo posts",
-                                    "Enables ability to undo posts before posting",
+                                    Utils.getResourceString("piko_pref_undo_posts"),
+                                    Utils.getResourceString("piko_pref_undo_posts_desc"),
                                     Settings.PREMIUM_UNDO_POSTS
                             )
                     );
 
                     premiumPrefs.addPreference(
                             buttonPreference(
-                                    "Undo posts settings",
+                                    Utils.getResourceString("piko_pref_undo_posts_btn"),
                                     "",
                                     Settings.PREMIUM_UNDO_POSTS.key
                             )
@@ -69,7 +86,7 @@ public class SettingsActivity extends Activity {
                 if (SettingsStatus.enableAppIconNNavIcon) {
                     premiumPrefs.addPreference(
                             buttonPreference(
-                                    "App icon and navbar customisation settings",
+                                    Utils.getResourceString("piko_pref_icon_n_navbar_btn"),
                                     "",
                                     Settings.PREMIUM_ICONS
                             )
@@ -78,26 +95,26 @@ public class SettingsActivity extends Activity {
             }
 
             if (SettingsStatus.changeDownloadEnabled) {
-                LegacyTwitterPreferenceCategory downloadPrefs = preferenceCategory("Download", screen);
+                LegacyTwitterPreferenceCategory downloadPrefs = preferenceCategory(Utils.getResourceString("piko_title_download"), screen);
                 downloadPrefs.addPreference(listPreference(
-                        "Public folder",
-                        "The public folder to use for video downloads",
+                        Utils.getResourceString("piko_pref_download_path"),
+                        Utils.getResourceString("piko_pref_download_path_desc"),
                         Settings.VID_PUBLIC_FOLDER
                 ));
                 downloadPrefs.addPreference(editTextPreference(
-                        "Download subfolder",
-                        "The subfolder to download videos to ([PublicFolder]/[Subfolder])",
+                        Utils.getResourceString("piko_pref_download_folder"),
+                        Utils.getResourceString("piko_pref_download_folder_desc"),
                         Settings.VID_SUBFOLDER
                 ));
             }
 
-            if(SettingsStatus.enableAdsSection()){
-                LegacyTwitterPreferenceCategory adsPrefs = preferenceCategory("Ads", screen);
+            if (SettingsStatus.enableAdsSection()) {
+                LegacyTwitterPreferenceCategory adsPrefs = preferenceCategory(Utils.getResourceString("piko_title_ads"), screen);
 
                 if (SettingsStatus.hideAds) {
                     adsPrefs.addPreference(
                             switchPreference(
-                                    "Hide promoted posts",
+                                    Utils.getResourceString("piko_pref_hide_promoted_posts"),
                                     "",
                                     Settings.ADS_HIDE_PROMOTED_POSTS
                             )
@@ -107,7 +124,7 @@ public class SettingsActivity extends Activity {
                 if (SettingsStatus.hideGAds) {
                     adsPrefs.addPreference(
                             switchPreference(
-                                    "Hide google ads",
+                                    Utils.getResourceString("piko_pref_hide_g_ads"),
                                     "",
                                     Settings.ADS_HIDE_GOOGLE_ADS
                             )
@@ -116,17 +133,16 @@ public class SettingsActivity extends Activity {
                 if (SettingsStatus.hideWTF) {
                     adsPrefs.addPreference(
                             switchPreference(
-                                    "Hide \"Who to follow\" section",
+                                    Utils.getResourceString("piko_pref_wtf_section"),
                                     "",
                                     Settings.ADS_HIDE_WHO_TO_FOLLOW
                             )
                     );
                 }
-
-                if (SettingsStatus.hideCTS) {
+   if (SettingsStatus.hideCTS) {
                     adsPrefs.addPreference(
                             switchPreference(
-                                    "Hide \"Creators to subscribe\" section",
+                                    Utils.getResourceString("piko_pref_cts_section"),
                                     "",
                                     Settings.ADS_HIDE_CREATORS_TO_SUB
                             )
@@ -136,7 +152,7 @@ public class SettingsActivity extends Activity {
                 if (SettingsStatus.hideCTJ) {
                     adsPrefs.addPreference(
                             switchPreference(
-                                    "Hide \"Community to join\" section",
+                                    Utils.getResourceString("piko_pref_ctj_section"),
                                     "",
                                     Settings.ADS_HIDE_COMM_TO_JOIN
                             )
@@ -146,7 +162,7 @@ public class SettingsActivity extends Activity {
                 if (SettingsStatus.hideRBMK) {
                     adsPrefs.addPreference(
                             switchPreference(
-                                    "Hide \"Revist your bookmark\" section",
+                                    Utils.getResourceString("piko_pref_ryb_section"),
                                     "",
                                     Settings.ADS_HIDE_REVISIT_BMK
                             )
@@ -156,7 +172,7 @@ public class SettingsActivity extends Activity {
                 if (SettingsStatus.hideRPinnedPosts) {
                     adsPrefs.addPreference(
                             switchPreference(
-                                    "Hide \"Pinned posts by followers\" section",
+                                    Utils.getResourceString("piko_pref_pinned_posts_section"),
                                     "",
                                     Settings.ADS_HIDE_REVISIT_PINNED_POSTS
                             )
@@ -166,7 +182,7 @@ public class SettingsActivity extends Activity {
                 if (SettingsStatus.hideDetailedPosts) {
                     adsPrefs.addPreference(
                             switchPreference(
-                                    "Hide detailed posts (in replies)",
+                                    Utils.getResourceString("piko_pref_hide_detailed_posts"),
                                     "",
                                     Settings.ADS_HIDE_DETAILED_POSTS
                             )
@@ -176,7 +192,7 @@ public class SettingsActivity extends Activity {
                 if (SettingsStatus.hidePromotedTrend) {
                     adsPrefs.addPreference(
                             switchPreference(
-                                    "Hide promoted trends",
+                                    Utils.getResourceString("piko_pref_hide_trends"),
                                     "",
                                     Settings.ADS_HIDE_PROMOTED_TRENDS
                             )
@@ -184,15 +200,14 @@ public class SettingsActivity extends Activity {
                 }
 
 
-
             }
 
             if (SettingsStatus.enableMiscSection()) {
-                LegacyTwitterPreferenceCategory miscPrefs = preferenceCategory("Misc", screen);
+                LegacyTwitterPreferenceCategory miscPrefs = preferenceCategory(Utils.getResourceString("piko_title_misc"), screen);
                 if (SettingsStatus.enableFontMod) {
                     miscPrefs.addPreference(
                             switchPreference(
-                                    "Enable chirp font",
+                                    Utils.getResourceString("piko_pref_chirp_font"),
                                     "",
                                     Settings.MISC_FONT
                             )
@@ -201,7 +216,7 @@ public class SettingsActivity extends Activity {
                 if (SettingsStatus.hideFAB) {
                     miscPrefs.addPreference(
                             switchPreference(
-                                    "Hide floating action button",
+                                    Utils.getResourceString("piko_pref_hide_fab"),
                                     "",
                                     Settings.MISC_HIDE_FAB
                             )
@@ -210,7 +225,7 @@ public class SettingsActivity extends Activity {
                 if (SettingsStatus.hideFABBtns) {
                     miscPrefs.addPreference(
                             switchPreference(
-                                    "Hide floating action button menu",
+                                    Utils.getResourceString("piko_pref_hide_fab_menu"),
                                     "",
                                     Settings.MISC_HIDE_FAB_BTN
                             )
@@ -220,7 +235,7 @@ public class SettingsActivity extends Activity {
                 if (SettingsStatus.hideRecommendedUsers) {
                     miscPrefs.addPreference(
                             switchPreference(
-                                    "Hide recommended users",
+                                    Utils.getResourceString("piko_pref_rec_users"),
                                     "",
                                     Settings.MISC_HIDE_RECOMMENDED_USERS
                             )
@@ -230,7 +245,7 @@ public class SettingsActivity extends Activity {
                 if (SettingsStatus.hideCommunityNote) {
                     miscPrefs.addPreference(
                             switchPreference(
-                                    "Hide community notes",
+                                    Utils.getResourceString("piko_pref_comm_notes"),
                                     "",
                                     Settings.MISC_HIDE_COMM_NOTES
                             )
@@ -240,7 +255,7 @@ public class SettingsActivity extends Activity {
                 if (SettingsStatus.hideViewCount) {
                     miscPrefs.addPreference(
                             switchPreference(
-                                    "Hide view count",
+                                    Utils.getResourceString("piko_pref_hide_view_count"),
                                     "",
                                     Settings.MISC_HIDE_VIEW_COUNT
                             )
@@ -250,20 +265,30 @@ public class SettingsActivity extends Activity {
                 if (SettingsStatus.customSharingDomainEnabled) {
                     miscPrefs.addPreference(
                             editTextPreference(
-                                    "Custom sharing domain",
-                                    "The domain to use when sharing tweets",
+                                    Utils.getResourceString("piko_pref_custom_share_domain"),
+                                    Utils.getResourceString("piko_pref_custom_share_domain_desc"),
                                     Settings.CUSTOM_SHARING_DOMAIN
+                            )
+                    );
+                }
+
+                if (SettingsStatus.featureFlags) {
+                    miscPrefs.addPreference(
+                            buttonPreference(
+                                    "Feature flags",
+                                    "",
+                                    Settings.MISC_FEATURE_FLAGS.key
                             )
                     );
                 }
             }
 
             if (SettingsStatus.enableTimelineSection()) {
-                LegacyTwitterPreferenceCategory timelinePrefs = preferenceCategory("Timeline", screen);
+                LegacyTwitterPreferenceCategory timelinePrefs = preferenceCategory(Utils.getResourceString("piko_title_timeline"), screen);
                 if (SettingsStatus.hideForyou) {
                     timelinePrefs.addPreference(
                             switchPreference(
-                                    "Hide for you tab",
+                                    Utils.getResourceString("piko_pref_hide_for_you"),
                                     "",
                                     Settings.TIMELINE_HIDE_FORYOU
                             )
@@ -272,8 +297,8 @@ public class SettingsActivity extends Activity {
                 if (SettingsStatus.hideLiveThreads) {
                     timelinePrefs.addPreference(
                             switchPreference(
-                                    "Hide live threads",
-                                    "",
+                                    Utils.getResourceString("piko_pref_hide_live_threads"),
+                                    Utils.getResourceString("piko_pref_hide_live_threads_desc"),
                                     Settings.TIMELINE_HIDE_LIVETHREADS
                             )
                     );
@@ -281,8 +306,8 @@ public class SettingsActivity extends Activity {
                 if (SettingsStatus.hideBanner) {
                     timelinePrefs.addPreference(
                             switchPreference(
-                                    "Hide banner",
-                                    "",
+                                    Utils.getResourceString("piko_pref_hide_banner"),
+                                    Utils.getResourceString("piko_pref_hide_banner_desc"),
                                     Settings.TIMELINE_HIDE_BANNER
                             )
                     );
@@ -290,7 +315,7 @@ public class SettingsActivity extends Activity {
                 if (SettingsStatus.hideInlineBmk) {
                     timelinePrefs.addPreference(
                             switchPreference(
-                                    "Hide bookmark icon in timeline",
+                                    Utils.getResourceString("piko_pref_hide_bmk_timeline"),
                                     "",
                                     Settings.TIMELINE_HIDE_BMK_ICON
                             )
@@ -300,12 +325,23 @@ public class SettingsActivity extends Activity {
                 if (SettingsStatus.showPollResultsEnabled) {
                     timelinePrefs.addPreference(
                             switchPreference(
-                                    "Show poll results",
-                                    "View poll results without voting",
+                                    Utils.getResourceString("piko_pref_show_poll_result"),
+                                    Utils.getResourceString("piko_pref_show_poll_result_desc"),
                                     Settings.TIMELINE_SHOW_POLL_RESULTS
                             )
                     );
                 }
+
+                if (SettingsStatus.hideImmersivePlayer) {
+                    timelinePrefs.addPreference(
+                            switchPreference(
+                                    Utils.getResourceString("piko_pref_hide_immersive_player"),
+                                    Utils.getResourceString("piko_pref_hide_immersive_player_desc"),
+                                    Settings.TIMELINE_HIDE_IMMERSIVE_PLAYER
+                            )
+                    );
+                }
+
             }
 
             setPreferenceScreen(screen);
@@ -318,7 +354,7 @@ public class SettingsActivity extends Activity {
             preference.setSummary(summary);
             preference.setKey(setting.key);
             preference.setDefaultValue(setting.defaultValue);
-
+            setOnPreferenceChangeListener(preference);
             return preference;
         }
 
@@ -328,7 +364,7 @@ public class SettingsActivity extends Activity {
             preference.setSummary(summary);
             preference.setKey(setting.key);
             preference.setDefaultValue(setting.defaultValue);
-
+            setOnPreferenceChangeListener(preference);
             return preference;
         }
 
@@ -338,7 +374,6 @@ public class SettingsActivity extends Activity {
             preference.setTitle(title);
             preference.setSummary(summary);
             preference.setOnPreferenceClickListener(this);
-
             return preference;
         }
 
@@ -351,7 +386,7 @@ public class SettingsActivity extends Activity {
             preference.setDefaultValue(setting.defaultValue);
             preference.setEntries(new CharSequence[]{"Movies", "DCIM", "Pictures", "Download"});
             preference.setEntryValues(new CharSequence[]{"Movies", "DCIM", "Pictures", "Download"});
-
+            setOnPreferenceChangeListener(preference);
             return preference;
         }
 
@@ -365,14 +400,40 @@ public class SettingsActivity extends Activity {
         @Override
         public boolean onPreferenceClick(Preference preference) {
             String key = preference.getKey();
-            if(key.equals(Settings.PREMIUM_UNDO_POSTS.key.toString())){
+            if (key.equals(Settings.PREMIUM_UNDO_POSTS.key.toString())) {
                 app.revanced.integrations.twitter.Utils.startUndoPostActivity();
-            }
-            else if(key.equals(Settings.PREMIUM_ICONS)){
+            } else if (key.equals(Settings.PREMIUM_ICONS)) {
                 app.revanced.integrations.twitter.Utils.startAppIconNNavIconActivity();
+            } else if (key.equals(Settings.MISC_FEATURE_FLAGS.key)) {
+                getFragmentManager().beginTransaction().replace(Utils.getResourceIdentifier("fragment_container", "id"), new FeatureFlagsFragment()).addToBackStack(null).commit();
             }
             return true;
         }
+
+        private void setOnPreferenceChangeListener(Preference preference) {
+            String key = preference.getKey();
+            preference.setOnPreferenceChangeListener(new Preference.OnPreferenceChangeListener() {
+                @Override
+                public boolean onPreferenceChange(Preference preference, Object newValue) {
+                    try{
+                        if (newValue!=null){
+                            if(newValue.getClass() == Boolean.class){
+                                app.revanced.integrations.twitter.Utils.putBooleanPerf(key,(Boolean)newValue);
+                            }
+                            else if(newValue.getClass() == String.class){
+                                app.revanced.integrations.twitter.Utils.putStringPerf(key,(String)newValue);
+                            }
+                        }
+
+                    }catch (Exception ex){
+                        Utils.showToastShort(ex.toString());
+                    }
+                    return true;
+                }
+            });
+        }
+
+        //end
     }