+++ app/src/main/java/app/revanced/integrations/twitter/settings/SettingsStatus.java
@@ -15,6 +15,7 @@ public class SettingsStatus {
     public static boolean hideBanner = false;
     public static boolean hideForyou = false;
     public static boolean showPollResultsEnabled = false;
+    public static boolean featureFlags = false;
 
     public static boolean hideAds = false;
     public static boolean hideGAds = false;
@@ -29,10 +30,13 @@ public class SettingsStatus {
     public static boolean enableUndoPosts = false;
     public static boolean enableAppIconNNavIcon = false;
 
+    public static boolean hideImmersivePlayer = false;
+
 
     public static void enableDownloadFolder() { changeDownloadEnabled = true; }
     public static void enableCustomSharingDomain() { customSharingDomainEnabled = true; }
     public static void enableFont() { enableFontMod = true; }
+    public static void enableFeatureFlags() { featureFlags = true; }
 
     public static void hideRecommendedUsers() { hideRecommendedUsers = true; }
     public static void hideCommunityNotes() { hideCommunityNote = true; }
@@ -58,8 +62,9 @@ public class SettingsStatus {
     public static void enableReaderMode() { enableReaderMode = true; }
     public static void enableUndoPosts() { enableUndoPosts = true; }
     public static void enableAppIconNNavIcon() { enableAppIconNNavIcon = true; }
+    public static void hideImmersivePlayer() { hideImmersivePlayer = true; }
 
-    public static boolean enableTimelineSection(){ return (hideLiveThreads || hideBanner || hideForyou || hideInlineBmk || showPollResultsEnabled); }
+    public static boolean enableTimelineSection(){ return (hideLiveThreads || hideBanner || hideForyou || hideInlineBmk || showPollResultsEnabled || hideImmersivePlayer); }
     public static boolean enableMiscSection() { return (enableFontMod || hideRecommendedUsers || hideFAB || hideViewCount || hideCommunityNote || customSharingDomainEnabled || hideFABBtns); }
     public static boolean enableAdsSection() {return (hideAds|| hideGAds || hideWTF || hideCTS || hideCTJ || hideDetailedPosts || hideRBMK ||hidePromotedTrend); }